add ruptures
Hello,

We actively maintain a package (`ruptures`) that does change point detection on time series. It consists in finding the instants where a signal changes its behaviour; it can be a shift in the mean, scale or more complex changes in the probability distribution, frequency, etc.
That would be great if you could add it to your list.

Also, thanks for this list. Usually, users of `ruptures` segment their signals and then classify each segment (to do activity recognition for instance), and several listed packages could be used for this purpose. I will certainly try one or two and add an illustrative example in the docs.

Cheers